Create volley
I also tried to filter out just the volleyball drop-in, but I can't find the right select class. I keep on getting an error message. You don't have to write the entire code, I need some guidelines on what I can try so it clicks on the right registration button for the right event. Once it does in the next step, I want this code to run by itself at a certain time so it registers me automatically.